clean architecture wip
