+++ config/db/database.py
@@ -1,0 +1,8 @@
+import os
+
+
+DATABASE_URL = os.getenv(
+    "DATABASE_URL",
+    "postgresql://postgres:postgres@localhost/api"
+)
+
